requirements.txt
numpy
psycopg2-binary
scipy
matplotlib
shapely